LayoutTests/http/tests/security/contentSecurityPolicy/window-open-javascript-url-with-target-blocked-expected.txt

CONSOLE MESSAGE: Refused to execute a script because its hash or 'unsafe-inline' does not appear in the script-src directive of the Content Security Policy.
